feat: adiciona helpers para criação de app e mocks para testes
