+++ backend/testes/helpers/app.ts
@@ -1,0 +1,8 @@
+import express, { Router } from 'express';
+
+export function criarApp(router: Router) {
+  const app = express();
+  app.use(express.json());
+  app.use('/', router);
+  return app;
+}
