+++ backend/testes/rotas/documentos.test.ts
@@ -1,53 +1,27 @@
-import express from 'express';
 import request from 'supertest';
 
 jest.mock('../../src/modulos/documentos/repositorio');
 jest.mock('../../src/modulos/usuarios/repositorio');
 jest.mock('../../src/servicos/armazenamento');
-jest.mock('../../src/middleware/autenticacao', () => ({
-  autenticar: (req: express.Request, _res: express.Response, next: express.NextFunction) => {
-    (req as any).usuario = {
-      sub: 'estudante-id',
-      perfil: 'estudante',
-      email: 'est@test.com',
-      nome: 'Estudante Teste',
-    };
-    next();
-  },
-}));
-jest.mock('../../src/middleware/autorizacao', () => ({
-  exigirPerfil: () => (_req: express.Request, _res: express.Response, next: express.NextFunction) => next(),
-}));
+jest.mock('../../src/middleware/autenticacao', () =>
+  require('../helpers/mocks').criarModuloAutenticacao('estudante-id', 'estudante', 'est@test.com', 'Estudante Teste')
+);
+jest.mock('../../src/middleware/autorizacao', () =>
+  require('../helpers/mocks').moduloAutorizacao
+);
 
 import * as repositorioDoc from '../../src/modulos/documentos/repositorio';
 import * as repositorioUsr from '../../src/modulos/usuarios/repositorio';
 import * as armazenamento from '../../src/servicos/armazenamento';
 import router from '../../src/modulos/documentos/rotas';
+import { criarApp } from '../helpers/app';
+import { DOC_MOCK } from '../helpers/fixtures';
 
 const mockDoc = repositorioDoc as jest.Mocked<typeof repositorioDoc>;
 const mockUsr = repositorioUsr as jest.Mocked<typeof repositorioUsr>;
 const mockArm = armazenamento as jest.Mocked<typeof armazenamento>;
 
-const app = express();
-app.use(express.json());
-app.use('/', router);
-
-const DOC_MOCK: repositorioDoc.Documento = {
-  id: 'doc-1',
-  titulo: 'Estágio XYZ',
-  tipo: 'estagio',
-  carga_horaria: 40,
-  estudante_id: 'estudante-id',
-  curso_id: 'curso-1',
-  status: 'pendente',
-  coordenador_id: null,
-  nome_arquivo: 'estagio.pdf',
-  caminho_arquivo: 'documentos/estudante-id/estagio.pdf',
-  tamanho_arquivo: 1024,
-  mime_type: 'application/pdf',
-  criado_em: new Date(),
-  atualizado_em: new Date(),
-};
+const app = criarApp(router);
 
 describe('GET /', () => {
   beforeEach(() => jest.clearAllMocks());
