+++ backend/testes/unitarios/middleware/autorizacao.test.ts
@@ -1,12 +1,6 @@
-import { Request, Response, NextFunction } from 'express';
+import { Request, NextFunction } from 'express';
 import { exigirPerfil } from '../../../src/middleware/autorizacao';
-
-function mockRes(): Response {
-  return {
-    status: jest.fn().mockReturnThis(),
-    json: jest.fn().mockReturnThis(),
-  } as unknown as Response;
-}
+import { criarMockRes } from '../../helpers/mocks';
 
 function mockReqComUsuario(perfil?: string): Request {
   return { usuario: perfil ? { perfil } : undefined } as unknown as Request;
@@ -20,7 +14,7 @@ describe('exigirPerfil', () => {
   });
 
   it('retorna 401 quando req.usuario não está definido', () => {
-    const res = mockRes();
+    const res = criarMockRes();
     exigirPerfil('coordenador')(mockReqComUsuario(), res, next);
     expect(res.status).toHaveBeenCalledWith(401);
     expect(res.json).toHaveBeenCalledWith({ erro: 'Não autenticado' });
@@ -28,7 +22,7 @@ describe('exigirPerfil', () => {
   });
 
   it('retorna 403 quando o perfil não tem permissão', () => {
-    const res = mockRes();
+    const res = criarMockRes();
     exigirPerfil('coordenador', 'admin')(mockReqComUsuario('estudante'), res, next);
     expect(res.status).toHaveBeenCalledWith(403);
     expect(res.json).toHaveBeenCalledWith({ erro: 'Sem permissão para este recurso' });
@@ -36,17 +30,17 @@ describe('exigirPerfil', () => {
   });
 
   it('chama next() quando o perfil tem permissão exata', () => {
-    exigirPerfil('coordenador')(mockReqComUsuario('coordenador'), mockRes(), next);
+    exigirPerfil('coordenador')(mockReqComUsuario('coordenador'), criarMockRes(), next);
     expect(next).toHaveBeenCalled();
   });
 
   it('chama next() quando o perfil está na lista de perfis aceitos', () => {
-    exigirPerfil('coordenador', 'admin')(mockReqComUsuario('admin'), mockRes(), next);
+    exigirPerfil('coordenador', 'admin')(mockReqComUsuario('admin'), criarMockRes(), next);
     expect(next).toHaveBeenCalled();
   });
 
   it('chama next() para admin quando vários perfis são aceitos', () => {
-    exigirPerfil('estudante', 'coordenador', 'admin')(mockReqComUsuario('admin'), mockRes(), next);
+    exigirPerfil('estudante', 'coordenador', 'admin')(mockReqComUsuario('admin'), criarMockRes(), next);
     expect(next).toHaveBeenCalled();
   });
 });
